created CRUD operation on claim/client/insurancePolicy

+++ src/main/java/com/example/insurance/entity/Claim.java
@@ -17,7 +17,7 @@ public class Claim {
 
     @Id
     @GeneratedValue
-    private double claimNumber;
+    private int claimNumber;
     private String description;
     private String claimDate;
     private String claimStatus;
