+++ src/main/java/com/example/insurance/repository/InsuranceRepository.java
@@ -3,6 +3,6 @@ package com.example.insurance.repository;
 import com.example.insurance.entity.InsurancePolicy;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InsuranceRepository extends JpaRepository<InsurancePolicy, Double> {
+public interface InsuranceRepository extends JpaRepository<InsurancePolicy, Integer> {
 
 }
